requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
colorama==0.4.6
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.14.1
pydantic==2.6.4
pydantic-core==2.16.3
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
dj-database-url==2.1.0
psycopg2==2.9.9
